[REF] web: Hoot - unify color management
This commit centralizes color management of Hoot UI with console colors.
To do this, colors declaration and management had to be moved to
Hoot-DOM, so that the same helpers could be used to declare CSS
variables, as well as console colors for event dispatching or other
logging features.
