+++ addons/web/static/lib/hoot/ui/hoot_style.css
@@ -403,6 +403,9 @@ input[type="checkbox"]:checked {
 .bg-amber {
     background-color: var(--amber);
 }
+.bg-blue {
+    background-color: var(--blue);
+}
 .bg-cyan {
     background-color: var(--cyan);
 }
@@ -435,6 +438,9 @@ input[type="checkbox"]:checked {
 .bg-amber-900 {
     background-color: var(--amber-900);
 }
+.bg-blue-900 {
+    background-color: var(--blue-900);
+}
 .bg-cyan-900 {
     background-color: var(--cyan-900);
 }
@@ -461,6 +467,9 @@ input[type="checkbox"]:checked {
 
 .border-amber {
     border-color: var(--amber);
+}
+.border-blue {
+    border-color: var(--blue);
 }
 .border-cyan {
     border-color: var(--cyan);
@@ -682,6 +691,9 @@ input[type="checkbox"]:checked {
 
 .col-span-2 {
     grid-column: span 2 / span 2;
+}
+.grid-cols-3 {
+    grid-template-columns: repeat(3, minmax(0, 1fr));
 }
 .grid-rows-3 {
     grid-template-rows: repeat(3, minmax(0, 1fr));
@@ -1362,6 +1374,9 @@ input[type="checkbox"]:checked {
 .text-amber {
     color: var(--amber);
 }
+.text-blue {
+    color: var(--blue);
+}
 .text-cyan {
     color: var(--cyan);
 }
@@ -1391,6 +1406,11 @@ input[type="checkbox"]:checked {
 .hover\:text-amber:focus-within,
 .hover\:text-amber:hover {
     color: var(--amber);
+}
+.hover\:text-blue:active,
+.hover\:text-blue:focus-within,
+.hover\:text-blue:hover {
+    color: var(--blue);
 }
 .hover\:text-cyan:active,
 .hover\:text-cyan:focus-within,
